+++ simple-stack-example-mvp-fragments/src/main/java/com/zhuinden/simplestackdemoexamplefragments/util/FragmentStateChanger.java
@@ -29,8 +29,7 @@ public class FragmentStateChanger {
             fragmentTransaction.setCustomAnimations(R.anim.slide_in_from_left, R.anim.slide_out_to_right);
         }
 
-        for(Object _oldKey : stateChange.getPreviousState()) {
-            Key oldKey = (Key) _oldKey;
+        for(Key oldKey : stateChange.<Key>getPreviousState()) {
             Fragment fragment = fragmentManager.findFragmentByTag(oldKey.getFragmentTag());
             if(fragment != null) {
                 if(!stateChange.getNewState().contains(oldKey)) {
@@ -40,8 +39,7 @@ public class FragmentStateChanger {
                 }
             }
         }
-        for(Object _newKey : stateChange.getNewState()) {
-            Key newKey = (Key) _newKey;
+        for(Key newKey : stateChange.<Key>getNewState()) {
             Fragment fragment = fragmentManager.findFragmentByTag(newKey.getFragmentTag());
             if(newKey.equals(stateChange.topNewState())) {
                 if(fragment != null) {