+++ simple-stack-flow-masterdetail-fragments/src/main/java/com/example/stackmasterdetailfrag/MasterDetailFragmentStateChanger.java
@@ -44,9 +44,8 @@ public class MasterDetailFragmentStateChanger {
             removeFragment(fragmentTransaction, NoDetailsPath.create());
         }
 
-        for(Object _previousKey : stateChange.getPreviousState()) {
-            Path previousKey = (Path) _previousKey;
-            if(!stateChange.getNewState().contains(_previousKey)) {
+        for(Path previousKey : stateChange.<Path>getPreviousState()) {
+            if(!stateChange.getNewState().contains(previousKey)) {
                 removeFragment(fragmentTransaction, previousKey);
             } else {
                 if(!previousKey.equals(masterKey) && !previousKey.equals(detailKey)) {
@@ -69,8 +68,7 @@ public class MasterDetailFragmentStateChanger {
             }
         }
 
-        for(Object _newKey : stateChange.getNewState()) {
-            Path newKey = (Path) _newKey;
+        for(Path newKey : stateChange.<Path>getNewState()) {
             if(!newKey.equals(masterKey) && !newKey.equals(detailKey)) {
                 detachFragment(fragmentTransaction, newKey);
             }